Showcase multilingual dense retrieval with E5
This new notebook shows a simple walkthrough of using the E5
multilingual embedding model to query a mixed language corpus. It
follows the same level of abstraction and details as the other notebooks
in this directory, opting to perform all inference outside of the stack.
We use select passages from the MIRACL dataset in English and German.