Adds Docker configuration for Elasticsearch, Kibana and APM server
Our recent example-apps all use OpenTelemetry, and have Docker instructions in
common. This copies the pertinent parts to a separate directory which can be
used by others as needed, without confusion.

If at some point start-local adds support for APM Server or EDOT collector,
we may be able to remove this and still have a working setup for Elasticsearch,
Kibana and OTLP collection.

Signed-off-by: Adrian Cole <adrian.cole@elastic.co>